specs: import integer helpers
Import a set of integer helpers which will be shared by the
specification modules. The main content of this helper module is an
abstraction of foreign-endian integers, allowing them to be represented
in the type-system and thus be protected from accidental access.

The module also has an abstraction over primitive-integers. However,
this abstraction is not exposed and only used for internal
simplification.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>